requirements/base.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
django-admin-inline-paginator==0.4.0
django-admin-sortable2==2.2.1
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-model-utils==4.5.1
django-modeltranslation==0.19.5
django-rosetta==0.10.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pillow==10.3.0
polib==1.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
singlemodeladmin==0.9
sniffio==1.3.1
sqlparse==0.5.0
urllib3==2.2.2
uvicorn==0.30.1
wait-for-it==2.2.2
